Extract & reuse ProjectBeneficiaryTable for listing beneficiaries
