+++ src/components/ProjectBeneficiariesTab.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Tab } from '@material-ui/core';
-import { formatMessage, PublishedComponent } from '@openimis/fe-core';
+import { formatMessage } from '@openimis/fe-core';
 import { PROJECT_BENEFICIARIES_TAB_VALUE, BENEFIT_PLAN_TYPE } from '../constants';
-import ProjectBeneficiariesSearcher from './ProjectBeneficiariesSearcher';
+import ProjectBeneficiaryTable from './ProjectBeneficiaryTable';
 // import ProjectGroupBeneficiariesSearcher from './ProjectGroupBeneficiariesSearcher';
 // TODO (Wei): handle group beneficiaries
 
@@ -24,19 +24,18 @@ function ProjectBeneficiariesTabPanel({
   value, project,
 }) {
   return (
-    <PublishedComponent
-      pubRef="policyHolder.TabPanel"
-      module="socialProtection"
-      index={PROJECT_BENEFICIARIES_TAB_VALUE}
-      value={value}
-    >
-      {project.benefitPlan?.type === BENEFIT_PLAN_TYPE.INDIVIDUAL ? (
-        <ProjectBeneficiariesSearcher project={project} />
-      ) : (
-        ''
-        // <ProjectGroupBeneficiariesSearcher project={project} />
+    <div hidden={value !== PROJECT_BENEFICIARIES_TAB_VALUE}>
+      {value === PROJECT_BENEFICIARIES_TAB_VALUE && (
+        <div>
+          {project.benefitPlan?.type === BENEFIT_PLAN_TYPE.INDIVIDUAL ? (
+            <ProjectBeneficiaryTable project={project} />
+          ) : (
+            ''
+            // <ProjectGroupBeneficiariesSearcher project={project} />
+          )}
+        </div>
       )}
-    </PublishedComponent>
+    </div>
   );
 }
